+++ backend/pong/utils/vector2.py
@@ -1,0 +1,10 @@
+class Vector2:
+    def __init__(self, x, y):
+        self.x = x
+        self.y = y
+    
+    def to_dict(self):
+        return {
+            "x": self.x,
+            "y": self.y
+        }
